Synchronize client commands with latest Metals.
It seems we forgot to update these in the shift to kebab-case.
Also, "run doctor" was missing.

These are commands that metals expects the client to support, for
example "focus diagnostics" or "run doctor". Status bar messages
can have attached client commands so that for example if we update the
status bar with "Build mis-configuration" the user can click on it to
open the doctor.