Improve version upgrade logic
Before this PR we detected when the user was running and old version of
Metals, and proposed to upgrade. However the upgrade happened only in
the Global (aka User) configuration, meaning that a user that had an
outdated version set in the Workspace or Workspace Folder Configuration
would see the "Old version detected" message again.

This PR fixes by inspecting the chain of configurations and finding the
one that is setting the old version, and upgrading it.

E.g. if you had

Global: 0.4.0
Workspace: 0.4.0
WorkspaceFolder: not set

Now we correctly bring you to

Global: 0.4.0
Workspace: 0.5.0
WorkspaceFolder: not set

If you then open another workspace where there is no specific setting:

Global: 0.4.0
Workspace: not set
WorkspaceFolder: not set

we then propose to upgrade again and bring you to

Global: 0.5.0
Workspace: not set
WorkspaceFolder: not set