Use string instead of array of string for `customRepositories`.
The benefit of string settings is that they can be edited directly in
the VS Code setting view. With `string[]` arrays you need to to edit the
raw `settings.json` file, which can be intimidating if you have never
done it before. We use space-separated values instead of pipe `|` to
for consistency with other settings like "server properties".