Implement Tree View Protocol (TVP).
This commit implements a new LSP extension called "Tree View Protocol"
which enables the language server to implement tree views such as:

- package explorer for library dependencies
- dynamically updating compilation progress overviews

The protocol is implemented as a thin shim on top of the VS Code tree
view API exposing the following functionality:

- List children: client->server request to go from a tree view node to
  its children nodes.
- Get parent: client->server request go from a tree view node to its
  parent node, if any.
- View visibility did change: client->server notification to notify the
  server that the visibility of a tree view changed.
- Node did change: server->client notification to notify the editor a
  tree view has updated.
- Reveal node: server->client notification to request the editor to
  focus on a tree view node.

Tree view nodes can have the following metadata:

- label: the title
- tooltip: shown on hover
- command: to execute when user clicks on the node
- icon: an svg to display next to the title
- collapse state: whether the node has children and if the node should auto-expand

The tree views are only started if the server supports the tree view
extensions. The server registers TVP support by adding the
`treeViewProvider: {}` field in the `initialize` LSP handshake.