Reveal the goto definition destination in the center of the editor view.
Previously, the cursor location after selecting goto definition for a
symbol in the tree view would not be centered if the symbol was already
in the view. This behavior made it sometimes difficult to find the
cursor location. Now the editor scrolls so that the cursor appears in
the center instead.