Add debugging provider to client capabilities.
Previously, it was not possible for the Metals server to know if the LSP
client supported debugging. This meant that "run" and "test" code lenses
appeared even for clients that don't support debugging.

Now, the VS Code client communicates to Metals that it supports
debugging so that Metals can disable code lenses for other clients.