Refine Java Home detection
First change: respect JAVA_HOME evironment variable, if defined.

The Metals website says that by default we try to use the `$JAVA_HOME`
environment variable while in reality we don't. This commit changes the
Java Home detection to work like it's explained in the documentation.

For people who start VS Code as a GUI application (not launched via
`code` command in the terminal) this change mostly likely won't have
an impact since environment variables that are defined in
`~/.bash_profile` and `~/.zshrc` are not set in GUI processes.

Second change: fallback to any Java 8 or Java 11 installation that's
available on the computer.  Previously, the VS Code extension failed to
launch on computers that only have Java 11 installed. Now, the VS Code
extension uses the most recent Java installation that's available that's
1.8 or higher and 1.11 or lower.

Sadly, the second fallback doesn't support Jabba installed JDKs but we
can maybe improve on that in the future.