Fix issues with hovers in worksheets
Previously we would send MarkupContent with `value` and `kind` parameter specified and feed it to decorations.
It seems that this would not work as it would not fit the expected format, which can be string, {language, string} or MarkdownString.
This should probably not worked before, but probably it was only by chance and as the behaviour was not documented it could be expected to last.