Offer to install Java in case of issues
Also, made a fix to only update the workspace config if that one was in fact invalid. This might be helpful in case of remote workspaces, where globally updating the config will break the local settings.

It would be ebst in those cases to update the remote setting, but it doesn't seem to be possible currently.