Fix issues with indentation when tying end markers
It turns out that we need to have the same number of indents as deindents, otherwise when we write `end m...` it will increase indentation n - 1, where n was the last indentation that was increased. In order to make it work properly we would also need to deindent without an end marker.

This changes the behaviour to only indent for onEnterRules, which does not work the same and does not require deindent.

I also added an additional rule so that no automatic indentation is made when we press enter before an expression.