Revert #717
As it turns out, we missed that in some cases Metals return incorrect
completions that were filtered out by VS Code.

For example:
```
// sample 1
for {
  x <- Some(1)
} y@@ // <- [Empty - scala.meta.internal.semanticdb.Tree, ..]

// sample 2
Input@@ // <- [InterruptedException]
```

Let's revert it and do release as the original issue was less annoying
than current state.

We can return this hack back in next release once this will be fixed on
server side.