Fix invoking focus on problems tab
Previously, each time an error was detected we would try to register a command indicated by `command` parameter if it was already registred. This would obviously fail as we explicitely checked for that. Now, whenever there is a problem in the workspace, the problem tab will be focused. Though the question is "should it?", wouldn't it be a bit annoying?