bugfix: Don't run command without clicking and fix infinite recursion
Previously, we would run a command without clicking on the status bar, which could be problematic if a message was shown multiple times. We would also invoke the RunDoctor command, which invokes itself. Now, RunDoctor invokes the correct part of the extension and status bar command is not invoked right away.

I am not sure why this never happened before, but might be a change in VS Code that prompted it.