improvement: Search PATH for Java to use
Previously, we would rely on the separate library for detecting java, but we had no control on whether it worked and it doesn't seem actively used by a lot of people.

Now, we first check JAVA_HOME and then search PATH for the actual Java to use. If we can't find any of that we download ourselves. This seems simple enought logic.