improvement: Set default Bloop version manually
Thanks to this, users will not get the pop up to update the version. I tried making the pop up just not appear, but if we run an old Metals instance at the same time, the other version will fall back to the old Bloop version causing an infinite loop.