[PLAT-145435] Hide sync element for unuspported files (#56)
The sync button sticks around after switching to an unsupported file. When clicking it, there's no effect and no user message, which is confusing. This becomes even more confusing now that python supports starts using the same UX, but creates the button from the databricks extension. Here we make sure we hide the button everytime we leave a supported editor