[PLAT-145435] Disable sync command on unsupported files (#57)
Today, the sync button is hidden if the server sets the status to hidden, but the command can still be run via the command palette. Here we completely disable the command if the server says so that that the server has full control on whether the sync is available or not.